pyproject.toml
[project]
name = "skrm-server"
version = "0.1.0"
description = "Backend API for the sKrm application"
authors = [{name = "Anthony Daga", email = "anthony@korrelated.com"}]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.121.2",
    "fido2>=2.0.0",
    "greenlet>=3.2.4",
    "langchain>=1.0.5",
    "langchain-openai>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "openai>=2.7.2",
    "pycrdt-websocket>=0.16.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "structlog>=25.5.0",
    "tenacity>=8.2.3",
    "uvicorn>=0.38.0",
    "uvloop>=0.22.1 ; sys_platform != 'win32'",
    "weaviate-client>=4.17.0,<5.0.0",
]

[dependency-groups]
test = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "httpx>=0.28.1",
    "aiosqlite>=0.21.0",
]
lint = [
    "black>=25.11.0",
    "isort>=5.12.0",
    "mypy>=1.18.2",
    "ruff>=0.14.5",
]
dev = [
    { include-group = "test" },
    { include-group = "lint" },
    "ipykernel>=7.1.0",
    "uv>=0.9.9",
    "jupyter>=1.1.1",
    "types-python-jose>=3.5.0.20250531",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "uvloop.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"app/models/*.py" = ["UP007", "UP037"]  # SQLAlchemy relationships need quoted forward references
"tests/**/*.py" = ["UP037"]  # Keep quotes for forward references in tests

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
concurrency = ["thread"]
source = ["app"]
omit = [
    "tests/**",
    "__init__.py",
    "app/core/weaviate/*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
